+++ libs/@guardian/atoms-rendering/src/YoutubeAtomOverlay.tsx
@@ -1,16 +1,16 @@
 import { css } from '@emotion/react';
-import type { ArticleTheme } from '@guardian/libs';
 import {
 	focusHalo,
-	neutral,
+	palette,
 	space,
 	textSans,
 } from '@guardian/source-foundations';
-import { SvgPlay } from '@guardian/source-react-components';
+import { SvgMediaControlsPlay } from '@guardian/source-react-components';
 import { formatTime } from './lib/formatTime';
-import { pillarPalette } from './lib/pillarPalette';
 import { Picture } from './Picture';
 import type { ImageSource, RoleType } from './types';
+
+export type VideoCategory = 'live' | 'documentary' | 'explainer';
 
 type Props = {
 	uniqueId: string;
@@ -21,9 +21,9 @@ type Props = {
 	alt: string;
 	role: RoleType;
 	duration?: number; // in seconds
-	pillar: ArticleTheme;
 	title?: string;
 	onClick: () => void;
+	videoCategory?: VideoCategory;
 };
 
 const overlayStyles = css`
@@ -60,11 +60,16 @@ const overlayStyles = css`
 	}
 `;
 
-const playButtonStyling = (pillar: ArticleTheme) => css`
-	background-color: ${pillarPalette[pillar][500]};
+const playButtonStyling = css`
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	margin-top: -40px; /* Half the height of the circle */
+	margin-left: -40px;
+	background-color: rgba(18, 18, 18, 0.6);
 	border-radius: 100%;
-	height: 60px;
-	width: 60px;
+	height: 80px;
+	width: 80px;
 	transform: scale(1);
 	transition-duration: 300ms;
 
@@ -73,26 +78,51 @@ const playButtonStyling = (pillar: ArticleTheme) => css`
 	justify-content: center;
 
 	svg {
-		fill: ${neutral[100]};
-		width: 45px;
-		height: 40px;
+		fill: ${palette.neutral[100]};
+		width: 75px;
+		height: 60px;
 	}
 `;
 
-const overlayInfoWrapperStyles = css`
-	display: flex;
-	flex-direction: row;
-	align-items: center;
+const pillStyles = css`
 	position: absolute;
-	bottom: ${space[4]}px;
-	left: ${space[4]}px;
+	top: ${space[2]}px;
+	right: ${space[2]}px;
+	${textSans.xxsmall({ fontWeight: 'bold' })};
+	background-color: rgba(0, 0, 0, 0.7);
+	color: ${palette.neutral[100]};
+	border-radius: ${space[3]}px;
+	padding: 0 6px;
+	display: inline-flex;
 `;
 
-const videoDurationStyles = (pillar: ArticleTheme) => css`
-	${textSans.medium({ fontWeight: 'bold' })};
-	padding-left: ${space[3]}px;
-	color: ${pillarPalette[pillar][500]};
+const pillItemStyles = css`
+	/* Target all but the first element, and add a border */
+	:nth-of-type(n + 2) {
+		border-left: 1px solid rgba(255, 255, 255, 0.5);
+	}
 `;
+
+const pillTextStyles = css`
+	line-height: ${space[4]}px;
+	padding: ${space[1]}px 6px;
+`;
+
+const liveStyles = css`
+	::before {
+		content: '';
+		width: 9px;
+		height: 9px;
+		border-radius: 50%;
+		background-color: ${palette.news[500]};
+		display: inline-block;
+		position: relative;
+		margin-right: 0.1875rem;
+	}
+`;
+
+const capitalise = (str: string): string =>
+	str.charAt(0).toUpperCase() + str.slice(1);
 
 export const YoutubeAtomOverlay = ({
 	uniqueId,
@@ -103,11 +133,14 @@ export const YoutubeAtomOverlay = ({
 	alt,
 	role,
 	duration,
-	pillar,
 	title,
 	onClick,
+	videoCategory,
 }: Props): JSX.Element => {
 	const id = `youtube-overlay-${uniqueId}`;
+	const hasDuration = duration !== undefined && duration > 0;
+	const showPill = !!videoCategory || hasDuration;
+	const isLive = videoCategory === 'live';
 	return (
 		<button
 			data-cy={id}
@@ -123,18 +156,29 @@ export const YoutubeAtomOverlay = ({
 				height={height}
 				width={width}
 			/>
-			<div css={overlayInfoWrapperStyles}>
-				<div
-					className="overlay-play-button"
-					css={css`
-						${playButtonStyling(pillar)}
-					`}
-				>
-					<SvgPlay />
+			{showPill && (
+				<div css={pillStyles}>
+					{!!videoCategory && (
+						<div css={pillItemStyles}>
+							<div css={[pillTextStyles, isLive && liveStyles]}>
+								{capitalise(videoCategory)}
+							</div>
+						</div>
+					)}
+					{!!hasDuration && (
+						<div css={pillItemStyles}>
+							<div css={pillTextStyles}>{formatTime(duration)}</div>
+						</div>
+					)}
 				</div>
-				{duration !== undefined && duration > 0 && (
-					<div css={videoDurationStyles(pillar)}>{formatTime(duration)}</div>
-				)}
+			)}
+			<div
+				className="overlay-play-button"
+				css={css`
+					${playButtonStyling}
+				`}
+			>
+				<SvgMediaControlsPlay />
 			</div>
 		</button>
 	);