+++ libs/@guardian/atoms-rendering/src/YoutubeAtomPlayer.tsx
@@ -1,10 +1,10 @@
 import type { Participations } from '@guardian/ab-core';
-import type { AdsConfig } from '@guardian/commercial-core';
+import type { AdsConfig } from '@guardian/commercial';
 import {
 	buildAdsConfigWithConsent,
 	buildImaAdTagUrl,
 	disabledAds,
-} from '@guardian/commercial-core';
+} from '@guardian/commercial';
 import type { ConsentState } from '@guardian/consent-management-platform/dist/types';
 import { log } from '@guardian/libs';
 import {
@@ -36,7 +36,7 @@ export declare class ImaManager {
 type Props = {
 	uniqueId: string;
 	videoId: string;
-	adTargeting?: AdTargeting;
+	adTargeting: AdTargeting;
 	consentState: ConsentState;
 	height: number;
 	width: number;