+++ libs/@guardian/libs/src/deprecated-exports.ts
@@ -1,0 +1,10 @@
+/**
+ * DEPRECATED EXPORTS
+ *
+ * To be removed from the public interface in the next major version.
+ */
+
+import { Pillar } from './format/Pillar';
+
+/** @deprecated Use `Pillar` instead. */
+export const ArticlePillar = Pillar;