+++ libs/@guardian/libs/src/format/ArticleDesign.ts
@@ -1,5 +1,6 @@
 export enum ArticleDesign {
 	Standard,
+	Picture,
 	Gallery,
 	Audio,
 	Video,