+++ tools/nx-plugins/eslint/executors/fix/executor.ts
@@ -1,5 +1,5 @@
-import type { ExecutorContext } from '@nrwl/devkit';
-import type { Schema } from '@nrwl/linter/src/executors/eslint/schema';
+import type { ExecutorContext } from '@nx/devkit';
+import type { Schema } from '@nx/linter/src/executors/eslint/schema';
 import linter from '../check/executor';
 
 // wrap @csnx/eslint:check executor with fix=true