+++ libs/@guardian/atoms-rendering/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @guardian/atoms-rendering
 
+## 32.2.1
+
+### Patch Changes
+
+- 1e15445: Pass valid ad targeting to YoutubeAtom stories
+
+## 32.2.0
+
+### Minor Changes
+
+- 61657a6: Delay initialisation of the Youtube player until we have ad targeting in place.
+
+## 32.1.0
+
+### Minor Changes
+
+- 7d933a5: Update the YoutubeAtom's duration pill to include the videoCategory, if provided
+- 4fafa9e: Updates the design of the youtube atom overlay
+
+## 32.0.1
+
+### Patch Changes
+
+- Update build output config
+
+## 32.0.0
+
+### Major Changes
+
+- a83c7a8: Update TS to v5.1.3 and all other packages to latest versions
+
+## 31.0.0
+
+### Major Changes
+
+- dbbbc58: Switch atoms dep from commercial-core to commercial and bump cmp
+
 ## 30.0.0
 
 ### Major Changes