+++ libs/@guardian/browserslist-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guardian/browserslist-config
 
+## 5.0.0
+
+### Major Changes
+
+- Update Typescript to v5.1.3 and tslib to v2.5.3
+
+## 4.2.0
+
+### Minor Changes
+
+- 3e6d282: Updated with June 2023 data
+
 ## 4.1.0
 
 ### Minor Changes