+++ libs/@guardian/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guardian/eslint-config
 
+## 4.1.0
+
+### Minor Changes
+
+- a135875: Bump eslint-config-prettier from 8.5.0 to 8.8.0
+
+## 4.0.0
+
+### Major Changes
+
+- Update Typescript to v5.1.3 and tslib to v2.5.3
+
 ## 3.0.0
 
 ### Major Changes